requirements.txt
requests
simplejson
pypiwin32
python-binance
scipy
PyJWT
